Refactored JedisServiceImplTest, now class using containers
