+++ src/main/java/com/example/app/config/JedisConfig.java
@@ -4,12 +4,13 @@ import org.springframework.boot.autoconfigure.data.redis.RedisProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import redis.clients.jedis.Jedis;
+import redis.clients.jedis.JedisPoolConfig;
 
 @Configuration
 public class JedisConfig {
 
     @Bean
-    public Jedis jedis(RedisProperties redisProperties) {
+    public Jedis jedisPoolConfig(RedisProperties redisProperties) {
         return new Jedis(redisProperties.getHost(), redisProperties.getPort());
     }
 }
