--- src/main/java/com/example/app/parser/service/JedisService.java
@@ -1,8 +1,0 @@
-package com.example.app.parser.service;
-
-public interface JedisService {
-
-    String saveKeyValue(String key, String value);
-
-    String getValueByKey(String key);
-}
