+++ src/main/java/com/example/app/service/JedisService.java
@@ -1,0 +1,7 @@
+package com.example.app.service;
+
+public interface JedisService {
+
+    String setValueByKey(String key, String value);
+    String getValueByKey(String key);
+}
